requirements.txt
apispec==6.7.1
astroid==3.2.4
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
coverage==7.6.4
cryptography==43.0.1
dill==0.3.8
execnet==2.1.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
flask-smorest==0.45.0
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mysqlclient==2.2.5
packaging==24.1
platformdirs==4.3.2
pluggy==1.5.0
psutil==6.1.0
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pylint==3.2.7
PyMySQL==1.1.1
pyotp==2.9.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.34
tomlkit==0.13.2
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.0.6
